ENGR00131573 SRTC Enable time set notification test
After upgrade to kernel version 2.6.35, the child process,
which called RTC_WAIT_TIME_SET ioctl, was not getting a chance
to execute until parent process, which changes the time, ran to
completion.

Therefore, it was necessary to swap the roles of the parent
and child processes so that parent calls RTC_WAIT_TIME_SET ioctl
while child changes the time. This ensures that the wait call
happens before the time change, a necessary condition for this
mechanism to work successfully.

Signed-off-by: Anish Trivedi <anish@freescale.com>