ENGR00155155 vpu: Change unit test to sync with v2.1.2 on mx6q
This change is based on v2.1.2 f/w release and corresponding
vpu API.

Note: Fileplay and prescan mode isn't supported from v2.1.2 on mx6q.
Normal mode and rollback mode can be supported in stream mode,
parameter "-s" is for the mode selection on i.mx6 platform
in the test.

Signed-off-by: Sammy He <r62914@freescale.com>