ENGR00162747 fix asrc sample rate convert issue
The original code assume that when input file length reach 0, the output
length should reach 0 too. Because of precision of algorithm calculating
the output file length, when convert 96k to 88.2k, the output file
length does not reach 0 when input length is 0. So the delete queue operation
of output buffer is blocked when there is no input buffer setuped and a
timeout error happens.

The new code break out of the "while(info->data_length>=0)" before the delete
queue operation of output buffer to make sure every time a delete output
queue operation happens, there is a input buffer set already.

Signed-off-by: Chen Liangjun <b36089@freescale.com>