ENGR00218972 MX6, SECURE BOOT, Change to Dynamic HAB Data Allocation
The original secure boot implementation make a consumption that
u-boot.bin will not exceed 0x2F000. With this consumption, the hab data
is hard coded in linker script file to relative address 0x2F000 without
causing any problem.

But when this consumption don't hold, the hard coded way will cause
memory region overlap and break build. So we need to change to a dynamic
way of allocating hab_data. The new implementation put hab data at the
next 0x1000 alignment after u-boot data and text section, instead of
hard coded to 0x2F000.

Similar changes is made to uImage authentication implementation. The new
secure image layout are as the following

U-Boot
    +-------------------+      DDR_START
    |                   |
    |    U-Boot Image   |
    |                   |
    +-------------------+      DDR_START + UBOOT_SIZE
    |    PADDING        |
    +-------------------+      align to 0x1000
    |    CSF Data       | \
    +-------------------+  +-- CSF + Pad,  Size : 0x2000
    |    PADDING        | /
    +-------------------+

uImage

    +-------------------+      DDR_START
    |                   |
    |     uImage        |
    |                   |
    +-------------------+      DDR_START + UIMAGE_SIZE
    |     PADDING       |
    +-------------------+      align to 0x1000
    |     IVT           | ---- Size : 0x20
    +-------------------+
    |     CSF Data      | \
    +-------------------+  +-- CSF + Pad,  Size : 0x2000
    |     PADDING       | /
    +-------------------+

The change will cause a lot of code signing script and configuration
file changes. For easy-to-use, they are covered by a all-in-one script.
For U-boot, the script is "mk_secure_uboot", for uImage, the script is
"mk_secure_uImage". They both use a 2-phase mechanism, first the CST
scripts and configuration files are generated using a file template,
second the generated script is excecuted to generated the final signed
image. The intermediate files are all deleted at the end. Please see
README for details.

Signed-off-by: Eric Sun <jian.sun@freescale.com>