ENGR00320056 add the option for users to choose the v4l2 device
Now that the driver supports dual camera, the camera/tvin test should
be able to choose the proper device for the testing.

The patch also adds options for setting output format and for saving to file
for csi_v4l2_capture (camera). Currently the driver does not support outputing
contents of both camera and tvin (via VADC) into one framebuffer. We should
output one of them to the file (now only for camera) when do dual camera
testing.

Signed-off-by: Robby Cai <r63905@freescale.com>